+++ forui/lib/src/widgets/button/button_content.dart
@@ -23,10 +23,7 @@ class Content extends StatelessWidget {
       child: DefaultTextStyle.merge(
         style: enabled ? contentStyle.enabledTextStyle : contentStyle.disabledTextStyle,
         child: IconTheme(
-          data: IconThemeData(
-            color: enabled ? contentStyle.enabledIconColor : contentStyle.disabledIconColor,
-            size: contentStyle.iconSize,
-          ),
+          data: enabled ? contentStyle.enabledIconStyle : contentStyle.disabledIconStyle,
           child: Row(
             mainAxisAlignment: MainAxisAlignment.center,
             spacing: 10,
@@ -72,26 +69,21 @@ final class FButtonContentStyle with Diagnosticable, _$FButtonContentStyleFuncti
   @override
   final EdgeInsetsGeometry padding;
 
-  /// The icon's color when this button is enabled.
+  /// The icon's style when this button is enabled.
   @override
-  final Color enabledIconColor;
+  final IconThemeData enabledIconStyle;
 
-  /// The icon's color when this button is disabled.
+  /// The icon's style when this button is disabled.
   @override
-  final Color disabledIconColor;
-
-  /// The icon's size. Defaults to 20.
-  @override
-  final double iconSize;
+  final IconThemeData disabledIconStyle;
 
   /// Creates a [FButtonContentStyle].
   FButtonContentStyle({
     required this.enabledTextStyle,
     required this.disabledTextStyle,
-    required this.enabledIconColor,
-    required this.disabledIconColor,
+    required this.enabledIconStyle,
+    required this.disabledIconStyle,
     this.padding = const EdgeInsets.symmetric(horizontal: 16, vertical: 12.5),
-    this.iconSize = 20,
   });
 
   /// Creates a [FButtonContentStyle] that inherits its properties from the given [enabled] and [disabled].
@@ -99,8 +91,8 @@ final class FButtonContentStyle with Diagnosticable, _$FButtonContentStyleFuncti
     : this(
         enabledTextStyle: typography.base.copyWith(color: enabled, fontWeight: FontWeight.w500, height: 1),
         disabledTextStyle: typography.base.copyWith(color: disabled, fontWeight: FontWeight.w500, height: 1),
-        enabledIconColor: enabled,
-        disabledIconColor: disabled,
+        enabledIconStyle: IconThemeData(color: enabled, size: 20),
+        disabledIconStyle: IconThemeData(color: disabled, size: 20),
       );
 }