+++ forui/test/src/widgets/button/button_golden_test.dart
@@ -254,6 +254,44 @@ void main() {
             matchesGoldenFile('button/${theme.name}/$variant/icon-disabled.png'),
           );
         });
+
+        testWidgets('${theme.name} with enabled circular progress', (tester) async {
+          await tester.pumpWidget(
+            TestScaffold(
+              theme: theme.data,
+              child: FButton(
+                prefix: const FProgress.circularIcon(),
+                label: const Text('Loading'),
+                onPress: () {},
+                style: variant,
+              ),
+            ),
+          );
+
+          await expectLater(
+            find.byType(TestScaffold),
+            matchesGoldenFile('button/${theme.name}/$variant/circular-progress-enabled-button.png'),
+          );
+        });
+
+        testWidgets('${theme.name} with disabled circular progress', (tester) async {
+          await tester.pumpWidget(
+            TestScaffold(
+              theme: theme.data,
+              child: FButton(
+                prefix: const FProgress.circularIcon(),
+                label: const Text('Loading'),
+                onPress: null,
+                style: variant,
+              ),
+            ),
+          );
+
+          await expectLater(
+            find.byType(TestScaffold),
+            matchesGoldenFile('button/${theme.name}/$variant/circular-progress-disabled-button.png'),
+          );
+        });
       }
     }
   });